add check for empty database

+++ scanner.py
@@ -104,10 +104,12 @@ def sync_database(data):
             # Get existing filenames from database
             cur.execute("SELECT filename FROM updates")
             rows = cur.fetchall()
-            old_files = []
 
-            for row in rows:
-                old_files.append(row[0])
+            if len(rows) > 0:
+                old_files = []
+
+                for row in rows:
+                    old_files.append(row[0])
 
             # Insert/update file information to database
             for ota in data:
@@ -119,13 +121,15 @@ def sync_database(data):
                         api_level=VALUES(api_level), url=VALUES(url), changes=VALUES(changes)""",
                         (ota["filename"], ota["device"], ota["incremental"], ota["timestamp"], ota["md5sum"], ota["channel"], ota["api_level"], ota["url"], ota["changes"]))
                 # File still exists so remove from to-be-removed list
-                old_files.remove(ota["filename"])
+                if len(rows) > 0:
+                    old_files.remove(ota["filename"])
 
             # Clean up old files from database
-            for old_file in old_files:
-                if DEBUG:
-                    print("Removing " + old_file)
-                cur.execute("DELETE FROM updates WHERE filename = %s", (old_file,))
+            if len(rows) > 0:
+                for old_file in old_files:
+                    if DEBUG:
+                        print("Removing " + old_file)
+                    cur.execute("DELETE FROM updates WHERE filename = %s", (old_file,))
 
     except mdb.Error, e:
         print("Error %d: %s" % (e.args[0],e.args[1]))
